# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentLens API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/agents/{id}/runs": {
      "get": {
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "unscored",
                "scored"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "versionLabel": {
                            "type": "string"
                          },
                          "input": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "unscored",
                              "scored"
                            ]
                          },
                          "overallScore": {
                            "nullable": true,
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "versionLabel",
                          "input",
                          "status",
                          "overallScore",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "page": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "total": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "limit",
                    "total"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "versionLabel": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "string"
                  },
                  "output": {
                    "type": "string"
                  },
                  "context": {},
                  "metadata": {}
                },
                "required": [
                  "versionLabel",
                  "input",
                  "output"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "unscored",
                        "scored"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/": {
      "get": {
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "unscored",
                "scored"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "agentId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "agentName",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "versionLabel",
            "required": false
          },
          {
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "oldest",
                "newest"
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "agentId": {
                            "type": "string"
                          },
                          "agentName": {
                            "type": "string"
                          },
                          "versionLabel": {
                            "type": "string"
                          },
                          "input": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "unscored",
                              "scored"
                            ]
                          },
                          "overallScore": {
                            "nullable": true,
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "agentId",
                          "agentName",
                          "versionLabel",
                          "input",
                          "status",
                          "overallScore",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "page": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "total": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "limit",
                    "total"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "versionLabel": {
                      "type": "string"
                    },
                    "input": {
                      "type": "string"
                    },
                    "output": {
                      "type": "string"
                    },
                    "context": {},
                    "metadata": {},
                    "status": {
                      "type": "string",
                      "enum": [
                        "unscored",
                        "scored"
                      ]
                    },
                    "overallScore": {
                      "nullable": true,
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "criteria": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "weight": {
                            "type": "number"
                          },
                          "score": {
                            "nullable": true,
                            "type": "object",
                            "properties": {
                              "value": {
                                "type": "number"
                              },
                              "justification": {
                                "nullable": true,
                                "type": "string"
                              }
                            },
                            "required": [
                              "value",
                              "justification"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "description",
                          "weight",
                          "score"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "versionLabel",
                    "input",
                    "output",
                    "context",
                    "metadata",
                    "status",
                    "overallScore",
                    "createdAt",
                    "criteria"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}/scores": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scores": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "criterionId": {
                          "type": "string"
                        },
                        "value": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 5
                        },
                        "justification": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "criterionId",
                        "value"
                      ]
                    }
                  }
                },
                "required": [
                  "scores"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "unscored",
                        "scored"
                      ]
                    },
                    "overallScore": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "overallScore"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}